requirements.txt
appdirs==1.4.4
asgiref==3.2.10
black==20.8b1
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
coreapi==2.3.3
coreschema==0.0.4
Django==3.1.1
django-filter==2.3.0
django-fsm==2.7.0
django-grappelli==2.14.2
djangorestframework==3.11.1
drf-yasg==1.17.1
flake8==3.8.3
idna==2.10
importlib-metadata==1.7.0
inflection==0.5.1
itypes==1.2.0
Jinja2==2.11.2
MarkupSafe==1.1.1
mccabe==0.6.1
mypy-extensions==0.4.3
packaging==20.4
pathspec==0.8.0
psycopg2-binary==2.8.6
pycodestyle==2.6.0
pyflakes==2.2.0
pyparsing==2.4.7
pytz==2020.1
regex==2020.7.14
requests==2.24.0
ruamel.yaml==0.16.12
ruamel.yaml.clib==0.2.2
six==1.15.0
sqlparse==0.3.1
toml==0.10.1
typed-ast==1.4.1
typing-extensions==3.7.4.3
uritemplate==3.0.1
urllib3==1.25.10
zipp==3.1.0
